Fuse methods and properties related to replies
