+++ src/matrix/room/timeline/relations.js
@@ -29,20 +29,6 @@ export function createAnnotation(targetId, key) {
     };
 }
 
-export function createReply(targetId, msgtype, body, formattedBody) {
-    return {
-        msgtype,
-        body,
-        "format": "org.matrix.custom.html",
-        "formatted_body": formattedBody,
-        "m.relates_to": {
-            "m.in_reply_to": {
-                "event_id": targetId
-            }
-        }
-    };
-}
-
 export function getRelationTarget(relation) {
     return relation.event_id || relation["m.in_reply_to"]?.event_id
 }
